submitting progress on styling
