+++ client/src/Pages/Shop/Shop.css
@@ -3,3 +3,7 @@
   flex-wrap: wrap;
   justify-content: center;
 }
+
+.loading {
+  margin: auto;
+}
